+++ config/config.go
@@ -105,6 +105,8 @@ var allRules = append([]lint.Rule{
 	&rule.UnnecessaryFormatRule{},
 	&rule.UseFmtPrintRule{},
 	&rule.EnforceSwitchStyleRule{},
+	&rule.IdenticalIfElseIfBranchesRule{},
+	&rule.IdenticalSwitchBranchesRule{},
 }, defaultRules...)
 
 // allFormatters is a list of all available formatters to output the linting results.
